update
-fixed session token in shop service
-added suspension checks on order and shop service
-updated tests
-all action require valid session token
-afek the best

+++ workshop-project/Backend/src/main/java/Service/OrderService.java
@@ -96,7 +96,11 @@ public class OrderService {
                 throw new Exception("User not logged in");
             }
             int cartID = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
-            Guest guest = userRepository.getUserById(cartID); // Get the guest user by I
+            Guest guest = userRepository.getUserById(cartID); // Get the guest user by ID
+            if(guest.isSuspended()) {
+                throw new Exception("User is suspended");
+            }
+
 
             
             Set<Integer> shopIds = userItems.keySet(); // Get the set of shop IDs
@@ -152,6 +156,9 @@ public class OrderService {
             }
             int cartID = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
             Guest guest = userRepository.getUserById(cartID); // Get the guest user by I
+            if(guest.isSuspended()) {
+                throw new Exception("User is suspended");
+            }
             purchaseService.removeItemsFromCart(guest, userItems); // Save the updated cart to the repository
             
             
@@ -184,6 +191,9 @@ public class OrderService {
             }
             int cartID = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
             Guest guest = userRepository.getUserById(cartID); // Get the guest user by I
+            if(guest.isSuspended()) {
+                throw new Exception("User is suspended");
+            }
             
             List<Pair<Integer, Integer>> locksToAcquire = new ArrayList<>();
         
@@ -273,6 +283,9 @@ public class OrderService {
                 }
                 int cartID = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
                 Guest guest = userRepository.getUserById(cartID); // Get the guest user by ID
+                if(guest.isSuspended()) {
+                    throw new Exception("User is suspended");
+                }
                 Shop shop = shopRepository.getShopById(shopId); // Get the shop by ID
                 purchaseService.submitBidOffer(guest,shop ,itemID, offerPrice);
     
@@ -365,8 +378,12 @@ public class OrderService {
             }
             int userId = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
             Guest guest = userRepository.getUserById(userId); // Get the guest user by ID
+            if(guest.isSuspended()) {
+                throw new Exception("User is suspended");
+            }
+            Registered user = userRepository.getUserByName(guest.getUsername());
             Shop shop = shopRepository.getShopById(shopId); // Get the shop by ID
-            Order order = purchaseService.purchaseBidItem(guest,shop,bidId, orderRepository.getAllOrders().size(),payment, shipment);
+            Order order = purchaseService.purchaseBidItem(user,shop,bidId, orderRepository.getAllOrders().size(),payment, shipment);
             orderRepository.addOrder(order); // Save the order to the repository
             logger.info(() -> "Bid item purchased successfully for bid ID: " + bidId);
             return Response.ok();
@@ -384,6 +401,9 @@ public class OrderService {
                 }
                 int cartID = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
                 Guest guest = userRepository.getUserById(cartID); // Get the guest user by ID
+                if(guest.isSuspended()) {
+                    throw new Exception("User is suspended");
+                }
                 Registered user = userRepository.getUserByName(guest.getUsername());
                 Shop shop = shopRepository.getShopById(shopId); // Get the shop by ID
                 purchaseService.submitAuctionOffer(user,shop ,auctionID, offerPrice);
@@ -404,6 +424,9 @@ public class OrderService {
             }
             int userId = Integer.parseInt(authenticationAdapter.getUsername(sessionToken));
             Guest guest = userRepository.getUserById(userId); // Get the guest user by ID
+            if(guest.isSuspended()) {
+                throw new Exception("User is suspended");
+            }
             Registered registered = userRepository.getUserByName(guest.getUsername());
             Shop shop = shopRepository.getShopById(shopId); // Get the shop by ID
             Order order = purchaseService.purchaseAuctionItem(registered,shop,auctionID, orderRepository.getAllOrders().size(),payment, shipment);
